[dual-vm] fixed genesis vm2
