+++ genesis/Cargo.toml
@@ -25,11 +25,7 @@ thiserror = { workspace = true }
 
 # dependencies for move2
 starcoin-vm2-storage = { workspace = true }
-starcoin-vm2-transaction-builder = { workspace = true }
-starcoin-vm2-types = { workspace = true }
-starcoin-vm2-crypto = { workspace = true }
-starcoin-vm2-statedb = { workspace = true }
-starcoin-vm2-executor = { workspace = true }
+starcoin-vm2-genesis = { workspace = true }
 
 [features]
 default = []
