+++ vm2/genesis/Cargo.toml
@@ -11,11 +11,11 @@ version.workspace = true
 
 [dependencies]
 anyhow = { workspace = true }
-starcoin-vm2-vm-runtime = { workspace = true }
 starcoin-vm2-vm-types = { workspace = true }
 starcoin-vm2-types = { workspace = true }
 starcoin-vm2-crypto = { workspace = true }
 starcoin-vm2-executor = { workspace = true }
 starcoin-vm2-state-api = { workspace = true }
 starcoin-vm2-transaction-builder = { workspace = true }
-starcoin-config = { workspace = true }
+starcoin-vm2-statedb = { workspace = true }
+starcoin-vm2-storage = { workspace = true }
