+++ vm2/genesis/src/lib.rs
@@ -1,26 +1,32 @@
-use starcoin_config::genesis_config::ChainNetwork;
-use starcoin_vm2_crypto::ed25519::genesis_key_pair;
+// Copyright (c) The Starcoin Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+use starcoin_vm2_crypto::{ed25519::genesis_key_pair, HashValue};
 use starcoin_vm2_executor::{
     block_executor2::BlockExecutedData, executor2::do_execute_block_transactions,
 };
 use starcoin_vm2_state_api::ChainStateWriter;
-use starcoin_vm2_transaction_builder::build_stdlib_package as build_stdlib_package_2;
+use starcoin_vm2_statedb::ChainStateDB;
+use starcoin_vm2_storage::{storage::StorageInstance, Storage};
+use starcoin_vm2_transaction_builder::build_stdlib_package;
 use starcoin_vm2_types::{
     account_config::CORE_CODE_ADDRESS,
     error::{BlockExecutorError, ExecutorResult},
     vm_error::KeptVMStatus,
 };
-use starcoin_vm2_types::{
-    transaction::{
-        Package, RawUserTransaction, SignedUserTransaction, Transaction, TransactionInfo,
-        TransactionPayload, TransactionStatus,
-    },
-    write_set::WriteSet,
-};
-use starcoin_vm2_vm_types::StateView;
 
-pub fn build_genesis_transaction_with_package(
-    net: &ChainNetwork,
+use starcoin_vm2_vm_types::{
+    genesis_config::GenesisConfig,
+    transaction::TransactionInfo,
+    transaction::{Package, RawUserTransaction, SignedUserTransaction, TransactionPayload},
+    transaction::{Transaction, TransactionStatus},
+    write_set::WriteSet,
+    StateView,
+};
+use std::sync::Arc;
+
+fn build_genesis_transaction_with_package(
+    chain_id: u8,
     package: Package,
 ) -> anyhow::Result<SignedUserTransaction> {
     let txn = RawUserTransaction::new_with_default_gas_token(
@@ -30,20 +36,41 @@ pub fn build_genesis_transaction_with_package(
         0,
         0,
         1, // init to 1 to pass time check
-        net.chain_id().id().into(),
+        chain_id.into(),
     );
     let (genesis_private_key, genesis_public_key) = genesis_key_pair();
     let sign_txn = txn.sign(&genesis_private_key, genesis_public_key)?;
     Ok(sign_txn.into_inner())
 }
 
-pub fn build_genesis_transaction(net: &ChainNetwork) -> anyhow::Result<SignedUserTransaction> {
-    let package = build_stdlib_package_2(
-        net.chain_id().id().into(),
-        net.genesis_config2().as_ref().unwrap(),
-        None,
-    )?;
-    build_genesis_transaction_with_package(net, package)
+fn build_genesis_transaction(
+    chain_id: u8,
+    genesis_config: &GenesisConfig,
+) -> anyhow::Result<SignedUserTransaction> {
+    let package = build_stdlib_package(chain_id.into(), genesis_config, None)?;
+    build_genesis_transaction_with_package(chain_id, package)
+}
+
+pub fn build_and_execute_genesis_transaction(
+    chain_id: u8,
+    genesis_config: &GenesisConfig,
+) -> (SignedUserTransaction, HashValue) {
+    let user_txn = build_genesis_transaction(chain_id, genesis_config).unwrap();
+
+    let storage = Arc::new(Storage::new(StorageInstance::new_cache_instance()).unwrap());
+    let chain_state = ChainStateDB::new(storage.clone(), None);
+
+    let txn = Transaction::UserTransaction(user_txn.clone());
+
+    (
+        user_txn,
+        execute_genesis_transaction(&chain_state, txn)
+            .unwrap()
+            .txn_infos
+            .pop()
+            .expect("Execute output must exist.")
+            .id(),
+    )
 }
 
 pub fn execute_genesis_transaction<S: StateView + ChainStateWriter + Sync>(
